cobrautil: decouple from bake/build dependency chain
