+++ commands/install.go
@@ -4,7 +4,6 @@ import (
 	"os"
 
 	"github.com/docker/buildx/util/cobrautil"
-	"github.com/docker/buildx/util/cobrautil/completion"
 	"github.com/docker/cli/cli"
 	"github.com/docker/cli/cli/command"
 	"github.com/docker/cli/cli/config"
@@ -49,7 +48,7 @@ func installCmd(dockerCli command.Cli) *cobra.Command {
 		},
 		Deprecated:            "use 'docker buildx' directly, without relying on the 'docker builder' alias",
 		Hidden:                true,
-		ValidArgsFunction:     completion.Disable,
+		ValidArgsFunction:     cobrautil.DisableCompletion,
 		DisableFlagsInUseLine: true,
 	}
